requirements.txt
vosk
openai-whisper
torch
numpy
scipy
